requirements.txt
fastmcp==2.13.1
matplotlib==3.10.7
matplotlib-inline==0.1.7
numpy==1.26.4
scipy==1.15.1
sse-starlette==3.0.3
starlette==0.45.3
uvicorn==0.38.0
pydantic>=2.0.0
python-dotenv
scikit-learn>=1.5.0
pandas>=2.2.0
seaborn>=0.13.0
statsmodels>=0.14.0
scikit-learn-extra>=0.3.0
lifelines>=0.27.0
semopy>=2.3.11
umap-learn>=0.5.0
hdbscan>=0.8.0